requirements.txt
fastmcp>=2.11.2
python-dotenv>=1.1.1
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
openai-whisper>=20231117
transformers>=4.36.0
torch>=2.1.0
google-generativeai>=0.3.0
cryptography>=41.0.0
dateparser>=1.2.0
Pillow>=10.0.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
beautifulsoup4>=4.12.0
readabilipy>=0.3.0
markdownify>=1.1.0
httpx>=0.25.0
numpy>=1.24.0
scikit-learn>=1.3.0